requirements.txt
numpy
opencv
torch
up42-blockutils